+++ cached_proc_macro/src/once.rs
@@ -6,12 +6,6 @@ use quote::quote;
 use syn::spanned::Spanned;
 use syn::{parse_macro_input, Ident, ItemFn, ReturnType};
 
-#[derive(Debug, Default, FromMeta)]
-enum SyncWriteMode {
-    #[default]
-    Default,
-}
-
 #[derive(FromMeta)]
 struct OnceMacroArgs {
     #[darling(default)]
@@ -19,7 +13,7 @@ struct OnceMacroArgs {
     #[darling(default)]
     time: Option<u64>,
     #[darling(default)]
-    sync_writes: Option<SyncWriteMode>,
+    sync_writes: bool,
     #[darling(default)]
     result: bool,
     #[darling(default)]
@@ -226,8 +220,8 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
         }
     };
 
-    let do_set_return_block = match args.sync_writes {
-        Some(SyncWriteMode::Default) => quote! {
+    let do_set_return_block = if args.sync_writes {
+        quote! {
             #r_lock_return_cache_block
             #w_lock
             if let Some(result) = &*cached {
@@ -235,13 +229,14 @@ pub fn once(args: TokenStream, input: TokenStream) -> TokenStream {
             }
             #function_call
             #set_cache_and_return
-        },
-        None => quote! {
+        }
+    } else {
+        quote! {
             #r_lock_return_cache_block
             #function_call
             #w_lock
             #set_cache_and_return
-        },
+        }
     };
 
     let signature_no_muts = get_mut_signature(signature);